requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
email-validator
pytest
pytest-cov
mutmut
typing-extensions